BoardRepository 하나로 Repository 정리
Jdbc Repository 삭제
CommentRepository 삭제
BoardRepository 인터페이스 삭제
JpaBoardRepository -> BoardRepository 이름 변경
