+++ src/main/java/hello/kssoftware/board/repository/BoardRepository.java
@@ -1,17 +1,51 @@
 package hello.kssoftware.board.repository;
 
-import hello.kssoftware.board.Board;
+import hello.kssoftware.board.entity.Board;
 import hello.kssoftware.board.dto.BoardSearchDto;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+import org.springframework.util.StringUtils;
 
 import java.util.List;
 
-public interface BoardRepository {
-    Long save(Board board);
+@Repository
+@RequiredArgsConstructor
+public class BoardRepository{
 
-    Board findById(Long id);
+    private final EntityManager em;
 
-    void delete(Long id);
+    public Long save(Board board) {
+        em.persist(board);
+        return board.getId();
+    }
 
-    List<Board> findAll(BoardSearchDto boardSearchDto);
+    public Board findById(Long id) {
+        return em.find(Board.class, id);
+    }
 
+    public void delete(Long id) {
+        Board findBoard = em.find(Board.class, id);
+        em.remove(findBoard);
+    }
+
+    public List<Board> findAll(BoardSearchDto boardSearchDto) {
+
+        String jpql = "select b From Board b left join fetch b.writer w left join fetch b.comments c ";
+        boolean isFirstCondition = true;
+
+        //제목으로 검색
+        if (StringUtils.hasText(boardSearchDto.getTitle())) {
+            jpql += " where b.title like '%" + boardSearchDto.getTitle() + "%'";
+            isFirstCondition = false;
+        }
+        //작성자로 검색
+        if (StringUtils.hasText(boardSearchDto.getWriterName())) {
+            jpql += isFirstCondition ? " where" : " and";
+            jpql += " b.writer.name like '%" + boardSearchDto.getWriterName() + "%'";
+        }
+        TypedQuery<Board> query = em.createQuery(jpql, Board.class).setMaxResults(1000); //최대 1000건
+        return query.getResultList();
+    }
 }
