--- src/main/java/hello/kssoftware/board/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package hello.kssoftware.board.repository;
-
-import hello.kssoftware.board.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
